Refactored models into seperate files (Now that we are more developers that would lead to fewer conflicts
