+++ api/index.js
@@ -4,7 +4,7 @@ const jwt = require('jsonwebtoken');
 
 const schema = require('./schema');
 const resolvers = require('./resolvers');
-const { getModels } = require('./database/models');
+const { getModels } = require('./database');
 const { getConnection } = require('./database/connection');
 
 const server = new ApolloServer({
